+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   of ObjSCIP without accessing the raw SCIP object:
   - message handlers derived from `scip::ObjMessagehdlr` via `Model::setMessagehdlr`,
   - problem data derived from `scip::ObjProbData` via a new constructor of `Model`,
+  - variable data derived from `scip::ObjVardata` via a new overload of `Model::addVar`,
   - Benders' decompositions derived from `scip::ObjBenders` via `Model::includeBenders`,
   - Benders' decomposition cuts derived from `scip::ObjBenderscut` via `Model::includeBenderscut`,
   - branching rules derived from `scip::ObjBranchrule` via `Model::includeBranchrule`,